26 Site Review and Tweak Finished!

+++ app/views/users/partials/follow-form.blade.php
@@ -1,11 +1,13 @@
-@if($user->isFollowedBy($currentUser))
-    {{ Form::open(['method' => 'DELETE', 'route' => ['unfollows_path', $user->id]] ) }}
-        {{ Form::hidden('userIdToUnfollow', $user->id) }}
-        <button type="submit" class="btn btn-danger">Unfollow {{ $user->username }}</button>
-    {{ Form::close() }}
-@else
-    {{ Form::open(['route' => 'follows_path']) }}
-        {{ Form::hidden('userIdToFollow', $user->id) }}
-        <button type="submit" class="btn btn-primary">Follow {{ $user->username }}</button>
-    {{ Form::close() }}
+@if($signedIn)
+    @if($user->isFollowedBy($currentUser))
+        {{ Form::open(['method' => 'DELETE', 'route' => ['unfollows_path', $user->id]] ) }}
+            {{ Form::hidden('userIdToUnfollow', $user->id) }}
+            <button type="submit" class="btn btn-danger">Unfollow {{ $user->username }}</button>
+        {{ Form::close() }}
+    @else
+        {{ Form::open(['route' => 'follows_path']) }}
+            {{ Form::hidden('userIdToFollow', $user->id) }}
+            <button type="submit" class="btn btn-primary">Follow {{ $user->username }}</button>
+        {{ Form::close() }}
+    @endif
 @endif
